Pill update & bug fixes
Pills are now actually drawn.
The player can no longer go through walls on more advanced maps.
The player can now only change direction (unless it's the opposite of the previous direction) when in the center of a tile.
added temporary MapParser class.
the background of the window is now black.
added placeholder images for the pill and super pill.